Add a custom DateTime type
Add a custom DateTime type that allows for timestamps to be used within
the database and `chrono::NaiveDateTime` to be used in the code. Diesel
will automatically convert between the two, so this can make it easier
to format the resulting representations and keep it consistent.

This does mean that sessions will have to be inserted with timestamps,
but `chrono` supports parsing timestamps from strings so this can be
somewhat automated.