Allow user votes to be recorded
Improve the voting page so that the submit button sends data to an API
endpoint which records the information in the database before
redirecting them to the page they came from. This will likely need a
`Flash<Message>` as currently it is too fast to see if anything has
happened.

Implement a custom struct for retrieving data from a form, allowing it
to be converted to Rust types but using a simple mapping, meaning it
doesn't require specific names of fields.