Enforce unique votes
On the voting page, users can enter the same candidate for all of their
rankings, thus allowing them to vote multiple times for the same person.
This makes a mess of the ranked choice voting system, as their first
vote may be ignored but replaced with the same person.

Instead, we should check whether their votes are unique before applying
them. If they are not, an error message will be displayed when they are
redirected, informing them that they must pick unique values.
Additionally, if their votes are counted they will now see a success
message telling them they have been entered to the database.

Closes #27.