Improve template hierarchy
Move more functionality to the `base` template, such as the 3 `div` tags
used for Bootstrap. This way, they can be removed from all other
templates and reduce indentation by 3 levels across the board.

Additionally, check whether there is a message to be displayed on every
template. While this reduces performance slightly on templates that do
not have a message to show, it makes it much easier to have consistent
messages across the website.