Only allow voting for open positions
When showing the list of positions that users can vote for, only allow
them to click on the open positions, showing the others as grayed out.
If users manage to find their way to the voting page for a closed
position, kick them back to the elections page with an error message.