feat: redirect on authentication
Remove the splash page that users would be placed on when they
authenticated and instead pass through the request URI to each handler.
When the user is authenticated correctly, they will then be redirected
to the original URI that they visited.

This makes it easier to navigate around the site, and means that a
direct link will always go to where it should do in the end.