feat: resolve ties using presidential ballot
When a tie occurs in the system (ie. there are more election winners
than were expected), use the ballot submitted by the president of the
club to determine the finalised winners. Ties in the ranking of
candidates are resolved by whoever came higher in the president's
ballot.

However, if Candidate A and B are tied for first place and Candidate C
is in third place with a presidential ballot of `[A, C, B]`, the result
will still be decided as `[A, B]` as B ranked higher in the overall vote
than C did.

Closes #42, #43.