feat: only elect when positions are closed
Previously, if an administrator viewed the results page during voting,
it would elect the candidate who is currently in the lead. This meant
that if users viewed the voting page after this event, they would not be
able to vote for that candidate.

Instead, only elect the winners of positions that are closed.

Closes #44.