refactor: improve access control mechanisms
Utilise the trait system and some generics to apply access control
measures at the routing level instead of inside handlers. This allows
for more expressive safety conditions and reduces the chance of a missed
check in the handler.

This also reduces the amount of code in the handler and moves
preconditions into the function signature itself.