feat: enable easier session creation
Add a page for managing sessions, which displays the upcoming sessions
and contains a form that allows people to create new sessions. Only site
administrators can view this page and send data to the endpoint, which
is defined by the `SITE_ADMINS` environment variable. Sessions are
created with a random identifier based on what is currently in the
database.

Closes #1.