feat: reduce sessions shown when managing
Reduce the number of sessions that are shown to users when managing them
by only showing those in the current window or the future. Previously,
it would just show all the sessions in the database, meaning users
needed to scroll down to the bottom of the list to find more recent
ones.

Closes #57.