feat: add migration for 2022 elections (#64)
With the Barbell elections approaching (3rd March) for 2022, we need to
update the database to support the new candidates and nominations. This
PR will drop the existing data in these tables (it has already been
backed up just in case) and adds the new mappings for who is running and
what for.

A separate change will be made on the infrastructure side before this is
merged, as this will update the election admins, site admins and who is
a member of Barbell before this change restarts all the Kubernetes pods.